Refactoring: - renamed renderDataDelegate -> renderTaskDelegate - moved renderTask to renderTaskDelegate
